fix(RedisShardBackplane): Don't return execute-only workers in getStorageWorkers (#1813)
* fix(RedisShardBackplane): Don't return execute-only workers in getStorageWorkers()

BuildFarm now registers all workers in RedisBackplane. When an execute-only worker is registered, RedisShardSubscriber will add it to RedisBackplane's storageWorkers. Then it's possible for getStorageWorkers() to reurn execute-only workers.
To fix this bug, We add a field worker_type to the proto message WorkerChange.Add so that RedisShardSubscriber can tell if the worker type matches the expected value.

* Cleans up code to address review comments